requirements.txt
locustio==0.11.0
beautifulsoup4==4.7.1
gevent==1.4.0
requests==2.21.0
har-transformer==1.2.4

